move changes to the use-binding-attributes file
