+++ packages/block-editor/src/hooks/index.js
@@ -28,8 +28,7 @@ import contentLockUI from './content-lock-ui';
 import './metadata';
 import blockHooks from './block-hooks';
 import blockRenaming from './block-renaming';
-import './block-binding-support';
-// import './use-bindings-attributes';
+import './use-bindings-attributes';
 
 createBlockEditFilter(
 	[
